server/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
DateTime==5.5
decorator==5.1.1
distlib==0.3.8
exceptiongroup==1.2.2
executing==2.0.1
Faker==28.0.0
filelock==3.13.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.5.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.1
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.32
SQLAlchemy-serializer==1.4.12
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.20.0
zope.interface==7.0.3
